requirements.txt
pdf2image
Pillow
python-docx
google-genai
python-dotenv
streamlit
